Module iui::controls

Available user interface controls and related functionality.

Note that Control and all specific control types are references to memory which is owned by the UI library.

Structs

Button
A textual button which users can click on, causing a callback to run.

Checkbox
Boolean selection control which can be checked or unchecked.

Combobox
Allows the user to select any one of its options, from a list shown only when selected.

Control
A generic UI control. Any UI control can be turned into this type.

Entry
Single-line editable text buffer.

Group
Collects controls together, with (optionally) a margin and/or title.

HorizontalBox
Lays out its children horizontally.

HorizontalSeparator
Horizontal line, to seperate things visually.

Label
A non-interactable piece of text.

LayoutGrid
Lays out its children in a grid according to insertion instructions.

MultilineEntry
Multi-line editable text buffer.

Slider
Numerical entry which allows users to select a value by picking a location along a line.

Spacer
Seperates components with empty space.

Spinbox
Numerical entry control which allows users to set any value in a range by typing or incrementing/decrementing.

TabGroup
Group of tabs, each of which shows a different sub-control.

VerticalBox
Lays out its children vertically.

Window
Contains a single child control and displays it and its children in a window on the screen.

Enums

GridAlignment
Informs a LayoutGrid how to align a control.

GridExpand
Informs a LayoutGrid about how a control should use available space in one or both dimensions.

GridInsertionStrategy
Informs a LayoutGrid as to position a control.

LayoutStrategy
Defines the ways in which the children of boxes can be layed out.

WindowType
A Window can either have a menubar or not; this enum represents that decision.\

Traits

NumericEntry

TextEntry