Struct iui::controls::Control

pub struct Control { /* fields omitted */ }

A generic UI control. Any UI control can be turned into this type.

Note that Control and all specific control types are references whose memory is owned by the UI library.

Methods

impl Control

pub unsafe fn from_ui_control(ui_control: *mut uiControl) -> Control

Creates a new Control object from an existing *mut uiControl.

pub fn as_ui_control(&self) -> *mut uiControl

Returns the underlying *mut uiControl.

pub unsafe fn destroy(&self)

Destroys a control. Any use of the control after this is use-after-free; therefore, this is marked unsafe.