Struct iui::controls::MultilineEntry

pub struct MultilineEntry { /* fields omitted */ }

Multi-line editable text buffer.

Methods

impl MultilineEntry

pub fn show(&mut self, _ctx: &UI)

pub fn hide(&mut self, _ctx: &UI)

pub unsafe fn from_raw(
uiMultilineEntry: *mut uiMultilineEntry
) -> MultilineEntry

Create an iui struct for this control from the raw pointer for it.

Unsafety

The given pointer must point to a valid control or memory unsafety may result.

pub fn ptr(&self) -> *mut uiMultilineEntry

Return the underlying pointer for this control.

impl MultilineEntry

pub fn new(_ctx: &UI) -> MultilineEntry