Module iui::draw

Functions and types related to 2D vector graphics.

Structs

BrushGradientStop

BrushRef
A reference to a DrawBrush

DrawContext
Drawing context, used to draw custom content on the screen.

LinearGradientBrush
A brush that paints a linear gradient.

Path

RadialGradientBrush
A brush that paints a radial gradient.

SolidBrush
A brush that paints all pixels with the same color, respecting alpha.

StrokeParams

StrokeParamsRef

Transform
A transformation which can be applied to the contents of a DrawContext.

Enums

Brush
Used to determine how a given stroke or fill is drawn.

FillMode

LineCap

LineJoin

Constants

DEFAULT_MITER_LIMIT